Tweak the lock screen to a Fairphone look'n'feel
1-  Left align the (big) clock and put the optional alarm above it.
2-  Implement a battery widget to meet the Blue Angel certification and inform
    when the phone is fully charged.
3-  The charging indication that was previously dispayed at the bottom of the
    lock screen now appears to the right of the battery widget.

The battery monitoring is realised locally even though we should be able to hook
up onto the BatteryController used in the status bar.
Also, there are a limited amount of battery assets. Another option would be to
follow the BatteryMeterView used in the status bar and *draw* the icon when
there is an event worth noticing.

FPIIM-202

Change-Id: I3d61ad41092e49be5352ea261735fe4606ac6dec